requirements.txt
cryptography==41.0.7
watchdog==4.0.0
pillow==10.3.0
pywin32==306
python-dateutil==2.8.2